error pages and broder for login\reg | change search users\add list users | corrects adaptive

+++ src/Components/Modal/Modal.js
@@ -1,12 +1,14 @@
 import classNames from 'classnames'
 import style from './Modal.module.css'
 import backDark from '../../img/back-dark.svg'
+import { useFetcher } from 'react-router-dom'
 
-export function Modal ({deleteUserState, setDeleteUserState, deleteAccount, submiteUpdates, setActiveModal, activeModal, passwodModal, setPasswordModal}){
+export function Modal ({deleteUserState, setDeleteUserState, deleteAccount, submiteUpdates, setActiveModal, activeModal, passwodModal, setPasswordModal,classErr, setClassErr}){
     return(
         <div className={activeModal || deleteUserState ? 'modal active-modal' : 'modal'} >
             <div onClick={(e) => {
                 e.preventDefault()
+                setClassErr('')
                 setDeleteUserState(false)
                 setActiveModal(false)
                 }} 
@@ -15,9 +17,12 @@ export function Modal ({deleteUserState, setDeleteUserState, deleteAccount, subm
             </div>
             <div className={style.containerModal}>
                 <p className={classNames(style.header, 'head-name')}>Enter Your Password</p>
-                <input className={classNames(style.input)} value={passwodModal} onChange={(e) => setPasswordModal(e.target.value)} type="password" />
+                <input className={classErr === '' ? classNames(style.input) : classNames(style.input, style.err)} value={passwodModal} onChange={(e) => setPasswordModal(e.target.value)} type="password" />
                 <button className={classNames(style.btn, 'head-name')} onClick={(event) => {
-                    if(deleteUserState){
+                    // if(classErr !== ''){
+                    //     setClassErr('error')
+                    // }
+                    if(deleteUserState && classErr === ''){
                         deleteAccount(event)
                         setDeleteUserState(false)
                         setActiveModal(false)
@@ -26,6 +31,7 @@ export function Modal ({deleteUserState, setDeleteUserState, deleteAccount, subm
                         submiteUpdates(event)
                         setActiveModal(false)
                     }
+                    
                     
                 }}>{deleteUserState ? "Delete" : "Log In"}</button>
             </div>
